ref(compiler): extract ReaderConditionalParser from Parser
Most of Parser's parse* methods already delegate to dedicated sub-parsers
via ExpressionParserFactory; reader conditionals were the exception, parsed
inline by two near-identical methods (parseReaderCondNode and
parseReaderCondSplicingNode) plus a private readNonTriviaExpression helper.

Move them into a ReaderConditionalParser sub-parser, created through the
factory like ListParser/QuoteParser/MetaParser. The shared branch scan and
closing-paren handling (duplicated verbatim between the two forms) collapse
into a single resolveBranch(); each public method only wraps the matched
branch into its node type.

Parser drops from 498 to ~388 LOC and stops carrying reader-conditional
state machinery. Parse tree output is unchanged: the parser unit/integration
suites (148) stay green, full compiler + core suites pass, and
bench_core_file_compilation is perf-neutral.
